+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added `Select Statement` action [#557](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/557)
 - Added `Remove Column` action [#556](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/556)
 - Added `Move Column Left` & `Move Column Right` actions [#562](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/562)
+- Added `Insert Column Left` & `Insert Column Right` actions [#563](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/563)
 - Added `Remove Table` action [#559](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/559)
 - Enabled removal of the UserRights block with `Remove Table` action [#560](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/560)
 - Ensure that Actions will respect readonly state of the files [#558](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/558)